Patch note #7  - added new methods for .java file;
 Message: now the server is fully processing all requests and works fine

+++ src/main/java/ru/harmony/serverboot/controller/SpecController.java
@@ -3,6 +3,7 @@ package ru.harmony.serverboot.controller;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import ru.harmony.serverboot.entity.Access;
 import ru.harmony.serverboot.entity.Spec;
 import ru.harmony.serverboot.entity.Worker;
 import ru.harmony.serverboot.response.BaseResponse;
@@ -18,10 +19,18 @@ public class SpecController {
 
     @GetMapping("/all")
     public ResponseEntity<ListResponse<Spec>> getAll() {
-        return ResponseEntity.ok(new ListResponse<Spec>(true, "Специализации", service.getAll()));
+        return ResponseEntity.ok(new ListResponse<Spec>(true, "Категории", service.getAll()));
     }
 
-    // place for getById method, if it needs
+    @GetMapping
+    // getById method
+    public ResponseEntity<DataResponse<Spec>> getById(@RequestParam Long id) {
+        if (service.findById(id).isPresent()) {
+            return ResponseEntity.ok(new DataResponse<Spec>(true, "Найдена категория", service.findById(id).orElseThrow()));
+        } else {
+            return ResponseEntity.badRequest().body(new DataResponse<Spec>(false, "Категория не найдена"));
+        }
+    }
 
     @PostMapping
     public ResponseEntity<DataResponse<Spec>> save(@RequestBody Spec spec) {
@@ -33,9 +42,12 @@ public class SpecController {
     public ResponseEntity<BaseResponse> deleteById(@RequestParam Long id) {
         // maybe try catch block
         // for check id in request body ? or not
-
-        service.delete(id);
-        return ResponseEntity.ok(new BaseResponse(true,"Категория была удален"));
+        if (service.findById(id).isPresent()) {
+            service.delete(id);
+            return ResponseEntity.ok(new BaseResponse(true, "Категория была удалена"));
+        } else {
+            return ResponseEntity.badRequest().body(new BaseResponse(false, "Категория не найдена и не удалена"));
+        }
     }
 
     @PutMapping
@@ -43,8 +55,11 @@ public class SpecController {
         // maybe try catch block
         // for check id in request body ? or not
         // and send message if it's true
-
-        service.update(spec);
-        return ResponseEntity.ok(new BaseResponse(true,"Данные категории были обновлены"));
+        if (service.findById(spec.getId()).isPresent()) {
+            service.update(spec);
+            return ResponseEntity.ok(new BaseResponse(true, "Данные категории были обновлены"));
+        } else {
+            return ResponseEntity.badRequest().body(new BaseResponse(false, "Категория не найдена и не удалена"));
+        }
     }
 }
