+++ src/main/java/ru/harmony/serverboot/controller/VacancyController.java
@@ -3,6 +3,7 @@ package ru.harmony.serverboot.controller;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import ru.harmony.serverboot.entity.Access;
 import ru.harmony.serverboot.entity.Spec;
 import ru.harmony.serverboot.entity.Vacancy;
 import ru.harmony.serverboot.response.BaseResponse;
@@ -22,6 +23,16 @@ public class VacancyController {
         return ResponseEntity.ok(new ListResponse<Vacancy>(true, "Вакансии", service.getAll()));
     }
 
+    @GetMapping
+    // getById method
+    public ResponseEntity<DataResponse<Vacancy>> getById(@RequestParam Long id) {
+        if (service.findById(id).isPresent()) {
+            return ResponseEntity.ok(new DataResponse<Vacancy>(true, "Найдена вакансия", service.findById(id).orElseThrow()));
+        } else {
+            return ResponseEntity.badRequest().body(new DataResponse<Vacancy>(false, "Вакансия не найдена"));
+        }
+    }
+
     @PostMapping
     public ResponseEntity<DataResponse<Vacancy>> save(@RequestBody Vacancy vacancy) {
         // maybe try catch block
@@ -30,11 +41,12 @@ public class VacancyController {
 
     @DeleteMapping
     public ResponseEntity<BaseResponse> deleteById(@RequestParam Long id) {
-        // maybe try catch block
-        // for check id in request body ? or not
-
-        service.delete(id);
-        return ResponseEntity.ok(new BaseResponse(true,"Вакансия удалена"));
+        if (service.findById(id).isPresent()) {
+            service.delete(id);
+            return ResponseEntity.ok(new BaseResponse(true, "Вакансия удалена"));
+        } else {
+            return ResponseEntity.badRequest().body(new BaseResponse(false, "Вакансия не найдена и не удалена"));
+        }
     }
 
     @PutMapping
@@ -42,8 +54,11 @@ public class VacancyController {
         // maybe try catch block
         // for check id in request body ? or not
         // and send message if it's true
-
-        service.update(vacancy);
-        return ResponseEntity.ok(new BaseResponse(true,"Вакансия была обновлена"));
+        if (service.findById(vacancy.getId()).isPresent()) {
+            service.update(vacancy);
+            return ResponseEntity.ok(new BaseResponse(true, "Вакансия была обновлена"));
+        } else {
+            return ResponseEntity.badRequest().body(new BaseResponse(false, "Вакансия не найдена и не удалена"));
+        }
     }
 }
